# docs/swagger.yaml
components:
  schemas:
    links.CreateLink:
      properties:
        description:
          type: string
        title:
          type: string
        url:
          type: string
      required:
      - title
      - url
      type: object
    user.AuthUser:
      properties:
        password:
          type: string
        username:
          type: string
      required:
      - password
      - username
      type: object
    user.CreateUser:
      properties:
        email:
          type: string
        name:
          type: string
        password:
          maxLength: 100
          minLength: 8
          type: string
        username:
          type: string
      required:
      - email
      - name
      - password
      - username
      type: object
    user.GetUser:
      properties:
        banner_picture:
          type: string
        bio:
          type: string
        created_at:
          type: string
        email:
          type: string
        id:
          type: integer
        name:
          type: string
        profile_picture:
          type: string
        username:
          type: string
      type: object
    user.UpdateBioRequest:
      properties:
        bio:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
externalDocs:
  description: ""
  url: ""
info:
  description: API do Linker, uma plataforma para gerenciamento de links e perfis
    personalizados.
  title: Linker API
  version: "1.0"
openapi: 3.1.0
paths:
  /profile/{username}:
    get:
      parameters:
      - description: username
        in: path
        name: username
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user.GetUser'
          description: OK
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Busca um perfil pelo nome de usuario
      tags:
      - profile
  /profile/banner:
    put:
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: file
        description: Imagem (jpg/png/webp)
        required: true
      responses:
        "204":
          content:
            application/json: {}
          description: No Content
        "401":
          content:
            application/json: {}
          description: Unauthorized
        "404":
          content:
            application/json: {}
          description: Not Found
        "422":
          content:
            application/json: {}
          description: Unprocessable Entity
        "500":
          content:
            application/json: {}
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Atualiza a foto de banner
      tags:
      - profile
  /profile/bio:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user.UpdateBioRequest'
        description: payload
        required: true
      responses:
        "204":
          content:
            application/json: {}
          description: No Content
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Atualiza a bio de um perfil
      tags:
      - profile
  /profile/link:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/links.CreateLink'
        description: payload
        required: true
      responses:
        "204":
          content:
            application/json: {}
          description: No Content
        "401":
          content:
            application/json: {}
          description: Unauthorized
        "404":
          content:
            application/json: {}
          description: Not Found
        "422":
          content:
            application/json: {}
          description: Unprocessable Entity
        "500":
          content:
            application/json: {}
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Cria um novo link para um usuário autenticado
      tags:
      - profile
  /profile/links/{username}:
    get:
      parameters:
      - description: username
        in: path
        name: username
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user.GetUser'
          description: OK
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Busca os links de um usuario
      tags:
      - profile
  /profile/photo:
    put:
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: file
        description: Imagem (jpg/png/webp)
        required: true
      responses:
        "204":
          content:
            application/json: {}
          description: No Content
        "401":
          content:
            application/json: {}
          description: Unauthorized
        "404":
          content:
            application/json: {}
          description: Not Found
        "422":
          content:
            application/json: {}
          description: Unprocessable Entity
        "500":
          content:
            application/json: {}
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Atualiza a foto de perfil
      tags:
      - profile
  /users:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user.CreateUser'
        description: payload
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: integer
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unprocessable Entity
        "500":
          content:
            application/json: {}
          description: Internal Server Error
      summary: Cria um novo usuario
      tags:
      - auth
  /users/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user.AuthUser'
        description: payload
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "422":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unprocessable Entity
        "500":
          content:
            application/json: {}
          description: Internal Server Error
      summary: autentica um usuário
      tags:
      - auth